model/requirements.txt
# Core libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
geopandas==0.13.2
shapely==2.0.1
pyproj==3.6.0
folium==0.14.0

# API and deployment
flask==2.3.2
gunicorn==21.2.0
flask-cors==4.0.0
flask-jwt-extended==4.5.2
python-dotenv==1.0.0

# Data processing
scipy==1.11.2
joblib==1.3.2

# For Arduino sensor integration
pyserial==3.5
